Upgrade Elastic4s, Play and Scala

+++ demo-ui/app/uk/gov/ons/addressIndex/demoui/views/ScoreHelper.scala
@@ -10,7 +10,7 @@ object ScoreHelper {
     * @param maxScore
     * @return
     */
-  def getPercentageFromScore(score: Float, maxScore: Float): Int = {
+  def getPercentageFromScore(score: Double, maxScore: Double): Int = {
     ((score / maxScore) * 100).toInt
   }
 
@@ -21,7 +21,7 @@ object ScoreHelper {
     * @param set
     * @return
     */
-  def getRelativePercentageScore(score: Float, set: AddressBulkResponseContainer): Int = {
+  def getRelativePercentageScore(score: Double, set: AddressBulkResponseContainer): Int = {
     ((score / set.bulkAddresses.maxBy(_.score).score) * 100).toInt
   }
